Removed some bugs

+++ src/libs/netlify/delete-dns.ts
@@ -1,4 +1,5 @@
 import { netlifyFetch } from "../netlify-fetch";
+import { listDNS } from "./list-dns";
 
 /** If response is undefined, then no errors occurred. Otherwise, there were errors. */
 type Response = undefined | {
@@ -9,8 +10,21 @@ type Response = undefined | {
    message: string,
 }
 
-export async function deleteDNS (dnsId: string) {
-	const data = await netlifyFetch<Response>(`dns_zones/sidharta_xyz/dns_records/${dnsId}`, {
+export async function deleteDNSById (dnsZone: string, dnsId: string) {
+	const data = await netlifyFetch<Response>(`dns_zones/${dnsZone}/dns_records/${dnsId}`, {
+		method: 'DELETE',
+	});
+	return data;
+}
+
+export async function deleteDNSByHostname (dnsZone: string, hostname: string) {
+	const recordsList = await listDNS(dnsZone);
+
+	const record = recordsList.find(record => record.hostname === hostname);
+
+	if (!record) throw new Error("Record not found on Netlify");
+
+	const data = await netlifyFetch<Response>(`dns_zones/${dnsZone}/dns_records/${record.id}`, {
 		method: 'DELETE',
 	});
 	return data;
